Выполянет 5.14. Модуляция
